Improved code style and readability. Major changes cover: - Renamed variables - Improved for loops structure and syntax - Set methods as private - Improved indentation - Renamed the folder
