+++ gsy_e_sdk/setups/grid_operator_api_scripts/redis_grid_fees.py
@@ -2,22 +2,22 @@
 # pylint: disable=duplicate-code
 
 """
-Template file for markets management through the gsy-e-sdk api client using Redis.
+Template file for markets management through the gsy-e-sdk api client using Rest.
 """
-import os
 from time import sleep
-from gsy_e_sdk.aggregator import Aggregator
-from gsy_e_sdk.rest_market import RestMarketClient
+from gsy_e_sdk.redis_aggregator import RedisAggregator
+from gsy_e_sdk.redis_market import RedisMarketClient
 from gsy_e_sdk.utils import log_grid_fees_information
-from gsy_e_sdk.utils import get_area_uuid_from_area_name_and_collaboration_id
 
-MARKET_NAMES = ["Grid", "Community"]
+MARKET_NAMES = [
+    "Grid",
+    "Community",
+]
 ORACLE_NAME = "dso"
 SLOT_LENGTH = 15  # leave as is
-AUTOMATIC = True
 
 
-class Oracle(Aggregator):
+class Oracle(RedisAggregator):
     """Class to represent the Grid Operator client type."""
 
     def __init__(self, *args, **kwargs):
@@ -32,10 +32,10 @@ class Oracle(Aggregator):
                 current_market_fee[area_dict["area_name"]] = area_dict[
                     "current_market_fee"
                 ]
-        next_market_fee = self.set_new_market_fee()
+        next_market_fee = self._set_new_market_fee()
         log_grid_fees_information(MARKET_NAMES, current_market_fee, next_market_fee)
 
-    def set_new_market_fee(self):
+    def _set_new_market_fee(self):
         """Return the market fees for each market for the next time slot."""
         next_market_fee = {}
         for area_uuid, area_dict in self.latest_grid_tree_flat.items():
@@ -56,50 +56,14 @@ class Oracle(Aggregator):
         self.is_finished = True
 
 
-def get_assets_name(node: dict) -> dict:
-    """
-    Parse the grid tree and return all registered assets
-    wrapper for _get_assets_name
-    """
-    if node == {}:
-        return {}
-    reg_assets = {"Area": [], "Load": [], "PV": [], "Storage": []}
-    _get_assets_name(node, reg_assets)
-    return reg_assets
-
-
-def _get_assets_name(node: dict, reg_assets: dict):
-    """
-    Parse the Collaboration / Canary Network registry
-    Return a list of the Market nodes the user is registered to
-    """
-    if node.get("registered") is True:
-        area_type = node["type"]
-        reg_assets[area_type].append(node["name"])
-    for child in node.get("children", []):
-        _get_assets_name(child, reg_assets)
-
-
-market_args = {
-    "simulation_id": os.environ["API_CLIENT_SIMULATION_ID"],
-    "domain_name": os.environ["API_CLIENT_DOMAIN_NAME"],
-    "websockets_domain_name": os.environ["API_CLIENT_WEBSOCKET_DOMAIN_NAME"],
-}
-
-aggregator = Oracle(aggregator_name=ORACLE_NAME, **market_args)
-if AUTOMATIC:
-    registry = aggregator.get_configuration_registry()
-    MARKET_NAMES = get_assets_name(registry)["Area"]
+MarketClient = RedisMarketClient
+aggregator = Oracle(aggregator_name=ORACLE_NAME)
 
 print()
 print("Connecting to markets ...")
 
 for i in MARKET_NAMES:
-    market_uuid = get_area_uuid_from_area_name_and_collaboration_id(
-        market_args["simulation_id"], i, market_args["domain_name"]
-    )
-    market_args["area_id"] = market_uuid
-    market_registered = RestMarketClient(**market_args)
+    market_registered = RedisMarketClient(area_id=i)
     market_registered.select_aggregator(aggregator.aggregator_uuid)
     print("----> Connected to ", i)
     sleep(0.3)
