+++ gsy_e_sdk/setups/grid_operator_api_scripts/resources/aziiz.json
