+++ gsy_e_sdk/setups/grid_operator_api_scripts/rest_aziiz_strategy.py
@@ -1,15 +1,18 @@
 # flake8: noqa
 # pylint: disable=duplicate-code
+
 """
 Template file to implement Aziiz grid fees
-strategy through the gsy-e-sdk api client using Redis.
+strategy through the gsy-e-sdk api client using Rest.
 """
 import os
 import json
 from time import sleep
-from gsy_e_sdk.redis_aggregator import RedisAggregator
-from gsy_e_sdk.redis_market import RedisMarketClient
+from gsy_e_sdk.aggregator import Aggregator
+from gsy_e_sdk.rest_market import RestMarketClient
 from gsy_e_sdk.utils import log_grid_fees_information
+from gsy_e_sdk.utils import get_area_uuid_from_area_name_and_collaboration_id
+
 
 module_dir = os.path.dirname(__file__)
 
@@ -20,10 +23,11 @@ MARKET_NAMES = [
 ORACLE_NAME = "dso"
 SLOT_LENGTH = 15
 MOVING_AVERAGE_PEAK = True
-LOOK_BACK = 4
+LOOK_BACK_INDEX = 4
+AUTOMATIC = True
 
 
-class Oracle(RedisAggregator):
+class Oracle(Aggregator):
     """Class that defines the behaviour of an "oracle" aggregator."""
 
     def __init__(self, *args, **kwargs):
@@ -44,7 +48,7 @@ class Oracle(RedisAggregator):
                     "current_market_fee"
                 ]
         self.calculate_import_export_balance()
-        next_market_fee = self.set_new_market_fee()
+        next_market_fee = self._set_new_market_fee()
         log_grid_fees_information(MARKET_NAMES, current_market_fee, next_market_fee)
 
     def calculate_import_export_balance(self):
@@ -63,7 +67,7 @@ class Oracle(RedisAggregator):
             )
         self.balance_hist.append(self.balance)
 
-    def set_new_market_fee(self):
+    def _set_new_market_fee(self):
         """Return the market fees for each market for the next time slot."""
         next_market_fee = {}
         for area_uuid, area_dict in self.latest_grid_tree_flat.items():
@@ -71,10 +75,10 @@ class Oracle(RedisAggregator):
                 if MOVING_AVERAGE_PEAK:
                     fees = []
                     # pylint: disable=unused-variable
-                    for j,k in enumerate(self.balance_hist):
+                    for j, k in enumerate(self.balance_hist):
                         fees.append(k[area_dict["area_name"]])
                     max_ext_energy_kwh = abs(
-                        sum(fees[-LOOK_BACK:]) / len(fees[-LOOK_BACK:])
+                        sum(fees[-LOOK_BACK_INDEX:]) / len(fees[-LOOK_BACK_INDEX:])
                     )
                 else:
                     max_ext_energy_kwh = max(
@@ -83,15 +87,9 @@ class Oracle(RedisAggregator):
                     )
 
                 individual_market_fees = fee_strategy[area_dict["area_name"]]
-                # pylint: disable=unused-variable
-                for j, k in enumerate(individual_market_fees):
-                    if (
-                        max_ext_energy_kwh
-                        <= k["Import / Export threshold"]
-                    ):
-                        next_market_fee[
-                            area_dict["area_name"]
-                        ] = k["Grid fee"]
+                for market_fee in individual_market_fees:
+                    if max_ext_energy_kwh <= market_fee["Import / Export threshold"]:
+                        next_market_fee[area_dict["area_name"]] = market_fee["Grid fee"]
                         break
                     next_market_fee[area_dict["area_name"]] = 2000
                 self.add_to_batch_commands.grid_fees(
@@ -119,14 +117,52 @@ def read_fee_strategy():
     return aziiz_fee
 
 
-aggregator = Oracle(aggregator_name=ORACLE_NAME)
+def get_assets_name(node: dict) -> dict:
+    """
+    Parse the grid tree and return all registered assets
+    wrapper for _get_assets_name
+    """
+    if node == {}:
+        return {}
+    reg_assets = {"Area": [], "Load": [], "PV": [], "Storage": []}
+    _get_assets_name(node, reg_assets)
+    return reg_assets
+
+
+def _get_assets_name(node: dict, reg_assets: dict):
+    """
+    Parse the Collaboration / Canary Network registry
+    Return a list of the Market nodes the user is registered to
+    """
+    if node.get("registered") is True:
+        area_type = node["type"]
+        reg_assets[area_type].append(node["name"])
+    for child in node.get("children", []):
+        _get_assets_name(child, reg_assets)
+
+
+market_args = {
+    "simulation_id": os.environ["API_CLIENT_SIMULATION_ID"],
+    "domain_name": os.environ["API_CLIENT_DOMAIN_NAME"],
+    "websockets_domain_name": os.environ["API_CLIENT_WEBSOCKET_DOMAIN_NAME"],
+}
+
+aggregator = Oracle(aggregator_name=ORACLE_NAME, **market_args)
+if AUTOMATIC:
+    registry = aggregator.get_configuration_registry()
+    MARKET_NAMES = get_assets_name(registry)["Area"]
+
 fee_strategy = read_fee_strategy()
 
 print()
 print("Connecting to markets ...")
 
 for i in MARKET_NAMES:
-    market_registered = RedisMarketClient(area_id=i)
+    market_uuid = get_area_uuid_from_area_name_and_collaboration_id(
+        market_args["simulation_id"], i, market_args["domain_name"]
+    )
+    market_args["area_id"] = market_uuid
+    market_registered = RestMarketClient(**market_args)
     market_registered.select_aggregator(aggregator.aggregator_uuid)
     print("----> Connected to ", i)
     sleep(0.3)
