+++ gsy_e_sdk/setups/grid_operator_api_scripts/rest_grid_fees.py
@@ -2,25 +2,16 @@
 # pylint: disable=duplicate-code
 
 """
-Template file to implement Time of Use grid fees
-strategy through the gsy-e-sdk api client using Rest.
+Template file for markets management through the gsy-e-sdk api client using Redis.
 """
 import os
-import csv
 from time import sleep
-from pendulum import from_format, DateTime
-from gsy_framework.constants_limits import DATE_TIME_FORMAT, TIME_FORMAT_SECONDS
 from gsy_e_sdk.aggregator import Aggregator
 from gsy_e_sdk.rest_market import RestMarketClient
 from gsy_e_sdk.utils import log_grid_fees_information
 from gsy_e_sdk.utils import get_area_uuid_from_area_name_and_collaboration_id
 
-module_dir = os.path.dirname(__file__)
-
-MARKET_NAMES = [
-    "Grid",
-    "Community",
-]
+MARKET_NAMES = ["Grid", "Community"]
 ORACLE_NAME = "dso"
 SLOT_LENGTH = 15  # leave as is
 AUTOMATIC = True
@@ -41,19 +32,15 @@ class Oracle(Aggregator):
                 current_market_fee[area_dict["area_name"]] = area_dict[
                     "current_market_fee"
                 ]
-        self.execute_batch_commands()
-        next_market_fee = self.set_new_market_fee(market_info)
+        next_market_fee = self._set_new_market_fee()
         log_grid_fees_information(MARKET_NAMES, current_market_fee, next_market_fee)
 
-    def set_new_market_fee(self, market_info):
+    def _set_new_market_fee(self):
         """Return the market fees for each market for the next time slot."""
         next_market_fee = {}
-        market_time = from_format(market_info["market_slot"], DATE_TIME_FORMAT)
         for area_uuid, area_dict in self.latest_grid_tree_flat.items():
             if area_dict["area_name"] in MARKET_NAMES:
-                next_market_fee[
-                    area_dict["area_name"]
-                ] = calculate_next_slot_market_fee(market_time, area_dict["area_name"])
+                next_market_fee[area_dict["area_name"]] = 10
 
                 self.add_to_batch_commands.grid_fees(
                     area_uuid=area_uuid,
@@ -67,36 +54,6 @@ class Oracle(Aggregator):
 
     def on_finish(self, finish_info):
         self.is_finished = True
-
-
-def read_fee_strategy():
-    "Return a dictionary containing the Time of Use strategy loaded from the CSV input file."
-    with open(
-        os.path.join(module_dir, "resources/ToU.csv"), newline="", encoding="utf-8"
-    ) as csvfile:
-        csv_rows = csv.reader(csvfile, delimiter=" ", quotechar="|")
-        headers = next(csv_rows)[0].split(";")
-        market_indexes = {}
-        planned_fee = {}
-        for market_name in MARKET_NAMES:
-            market_indexes.update({(market_name, headers.index(market_name))})
-        for row in csv_rows:
-            row = row[0].split(";")
-            for market in MARKET_NAMES:
-                planned_fee.update({(row[0], market): row[market_indexes[market]]})
-    return planned_fee
-
-
-def calculate_next_slot_market_fee(market_time: DateTime, market_name:str) -> float:
-    """Return the market fee for the next time slot."""
-    slot_time = market_time.add(minutes=SLOT_LENGTH).format(TIME_FORMAT_SECONDS)
-    if (slot_time, market_name) in fee_strategy:
-        next_fee = fee_strategy[(slot_time, market_name)]
-        if not isinstance(next_fee, (int, float)):
-            next_fee = float(next_fee.replace(",", "."))
-    else:
-        next_fee = None
-    return next_fee
 
 
 def get_assets_name(node: dict) -> dict:
@@ -133,8 +90,6 @@ aggregator = Oracle(aggregator_name=ORACLE_NAME, **market_args)
 if AUTOMATIC:
     registry = aggregator.get_configuration_registry()
     MARKET_NAMES = get_assets_name(registry)["Area"]
-
-fee_strategy = read_fee_strategy()
 
 print()
 print("Connecting to markets ...")
