+++ gsy_e_sdk/setups/grid_operator_api_scripts/rest_tou_strategy.py
@@ -1,16 +1,19 @@
 # flake8: noqa
+# pylint: disable=duplicate-code
+
 """
 Template file to implement Time of Use grid fees
-strategy through the gsy-e-sdk api client using Redis.
+strategy through the gsy-e-sdk api client using Rest.
 """
 import os
 import csv
 from time import sleep
 from pendulum import from_format, DateTime
 from gsy_framework.constants_limits import DATE_TIME_FORMAT, TIME_FORMAT_SECONDS
-from gsy_e_sdk.redis_aggregator import RedisAggregator
-from gsy_e_sdk.redis_market import RedisMarketClient
+from gsy_e_sdk.aggregator import Aggregator
+from gsy_e_sdk.rest_market import RestMarketClient
 from gsy_e_sdk.utils import log_grid_fees_information
+from gsy_e_sdk.utils import get_area_uuid_from_area_name_and_collaboration_id
 
 module_dir = os.path.dirname(__file__)
 
@@ -20,9 +23,10 @@ MARKET_NAMES = [
 ]
 ORACLE_NAME = "dso"
 SLOT_LENGTH = 15  # leave as is
+AUTOMATIC = True
 
 
-class Oracle(RedisAggregator):
+class Oracle(Aggregator):
     """Class to represent the Grid Operator client type."""
 
     def __init__(self, *args, **kwargs):
@@ -37,10 +41,11 @@ class Oracle(RedisAggregator):
                 current_market_fee[area_dict["area_name"]] = area_dict[
                     "current_market_fee"
                 ]
-        next_market_fee = self.set_new_market_fee(market_info)
+        self.execute_batch_commands()
+        next_market_fee = self._set_new_market_fee(market_info)
         log_grid_fees_information(MARKET_NAMES, current_market_fee, next_market_fee)
 
-    def set_new_market_fee(self, market_info):
+    def _set_new_market_fee(self, market_info):
         """Return the market fees for each market for the next time slot."""
         next_market_fee = {}
         market_time = from_format(market_info["market_slot"], DATE_TIME_FORMAT)
@@ -67,7 +72,7 @@ class Oracle(RedisAggregator):
 def read_fee_strategy():
     "Return a dictionary containing the Time of Use strategy loaded from the CSV input file."
     with open(
-            os.path.join(module_dir, "resources/ToU.csv"), newline="", encoding="utf-8"
+        os.path.join(module_dir, "resources/ToU.csv"), newline="", encoding="utf-8"
     ) as csvfile:
         csv_rows = csv.reader(csvfile, delimiter=" ", quotechar="|")
         headers = next(csv_rows)[0].split(";")
@@ -94,14 +99,52 @@ def calculate_next_slot_market_fee(market_time: DateTime, market_name:str) -> fl
     return next_fee
 
 
-aggregator = Oracle(aggregator_name=ORACLE_NAME)
+def get_assets_name(node: dict) -> dict:
+    """
+    Parse the grid tree and return all registered assets
+    wrapper for _get_assets_name
+    """
+    if node == {}:
+        return {}
+    reg_assets = {"Area": [], "Load": [], "PV": [], "Storage": []}
+    _get_assets_name(node, reg_assets)
+    return reg_assets
+
+
+def _get_assets_name(node: dict, reg_assets: dict):
+    """
+    Parse the Collaboration / Canary Network registry
+    Return a list of the Market nodes the user is registered to
+    """
+    if node.get("registered") is True:
+        area_type = node["type"]
+        reg_assets[area_type].append(node["name"])
+    for child in node.get("children", []):
+        _get_assets_name(child, reg_assets)
+
+
+market_args = {
+    "simulation_id": os.environ["API_CLIENT_SIMULATION_ID"],
+    "domain_name": os.environ["API_CLIENT_DOMAIN_NAME"],
+    "websockets_domain_name": os.environ["API_CLIENT_WEBSOCKET_DOMAIN_NAME"],
+}
+
+aggregator = Oracle(aggregator_name=ORACLE_NAME, **market_args)
+if AUTOMATIC:
+    registry = aggregator.get_configuration_registry()
+    MARKET_NAMES = get_assets_name(registry)["Area"]
+
 fee_strategy = read_fee_strategy()
 
 print()
 print("Connecting to markets ...")
 
 for i in MARKET_NAMES:
-    market_registered = RedisMarketClient(area_id=i)
+    market_uuid = get_area_uuid_from_area_name_and_collaboration_id(
+        market_args["simulation_id"], i, market_args["domain_name"]
+    )
+    market_args["area_id"] = market_uuid
+    market_registered = RestMarketClient(**market_args)
     market_registered.select_aggregator(aggregator.aggregator_uuid)
     print("----> Connected to ", i)
     sleep(0.3)
